requirements.txt
tweepy[async]
python-dotenv
black
apscheduler
